Finally fixed the minimization problem!
Before, I just used the window's bakcbuffer to store the pixel data, but
that backbuffer is not supposed to be used for that. Furthermore,
nothing would be written to it if the window was minimized or if the
computer went to sleep or anything like that. Now I use a Surface to
store pixel data and copy it to a texture when I need to update the
screen.